fix: delete

+++ src/main/java/br/com/brunobrolesi/parking/service/VehicleService.java
@@ -41,6 +41,10 @@ public class VehicleService {
     }
 
     public void delete(Integer id) {
+        Optional<Vehicle> optional = repository.findById(id);
+
+        if (optional.isEmpty()) throw new IllegalArgumentException("Este id é invalido");
+
         repository.deleteById(id);
     }
